Create SoilMoisture_Code
A soil moisture sensor is used to measure the water content in soil. It works on the principle of electrical conductivity. When the soil is moist, it conducts electricity better; when dry, conductivity is low. The sensor output changes based on this moisture level.
It is commonly used in automatic watering systems, agriculture, and plant monitoring systems.

A typical soil moisture sensor module has four main pins:
1. VCC – Connects to the 5V power supply from Arduino
2. GND – Connects to the ground (GND) of Arduino
3. AO (Analog Output) – Sends the analog moisture value to Arduino
4. DO (Digital Output) – Sends HIGH or LOW based on a threshold set using the onboard potentiometer

Working Principle
1. Dry soil → High resistance → Higher sensor value
2. Wet soil → Low resistance → Lower sensor value
3. The Arduino reads this value from the AO pin and displays it on the Serial Monitor.
4. If DO is used, the sensor gives HIGH/LOW output based on a threshold set by the blue potentiometer on the module.

Applications
>Automatic irrigation systems
>Smart gardening
>Greenhouse monitoring
>Agricultural projects